requirements.txt
# Core Dependencies
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# Audio Processing
pyaudio>=0.2.13
librosa>=0.10.0
soundfile>=0.12.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0  # Progress bars
